requirements.txt

numpy
scipy
networkx
matplotlib
sentence-transformers
torch
faiss-cpu
pandas
pyyaml
flask
gradio
scikit-learn
soundfile
librosa
